+++ lib/ui/settings2/Settings2GovernmentIdPanel.dart
@@ -108,8 +108,10 @@ class _Settings2GovernmentIdPanelPanelState extends State<Settings2GovernmentIdP
 
     await Health().deleteUser();
     await Exposure().deleteUser();
-    await Auth().deleteUserPiiData();
-    await User().deleteUser();
+    bool piiDeleted = await Auth().deleteUserPiiData();
+    if(piiDeleted) {
+      await User().deleteUser();
+    }
     Auth().logout();
   }
